+++ aclk/mqtt.c
@@ -7,34 +7,33 @@
 
 void (*_on_connect)(void *ptr) = NULL;
 void (*_on_disconnect)(void *ptr) = NULL;
-extern int cmdpause;
-
 
 #ifndef ENABLE_ACLK
 
 inline const char *_link_strerror(int rc)
 {
-    (void) rc;
+    UNUSED(rc);
     return "no error";
 }
 
 int _link_event_loop(int timeout)
 {
-    (void) timeout;
+    UNUSED(timeout);
     return 0;
 }
 
-int _link_send_message(char *topic, char *message)
+int _link_send_message(char *topic, char *message, int *mid)
 {
-    (void) topic;
-    (void) message;
+    UNUSED(topic);
+    UNUSED(message);
+    UNUSED(mid);
     return 0;
 }
 
 int _link_subscribe(char  *topic, int qos)
 {
-    (void) topic;
-    (void) qos;
+    UNUSED(topic);
+    UNUSED(qos);
     return 0;
 }
 
@@ -45,67 +44,30 @@ void _link_shutdown()
 
 int _link_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *), void (*on_disconnect)(void *))
 {
-    (void) aclk_hostname;
-    (void) aclk_port;
-    (void) on_connect;
-    (void) on_disconnect;
+    UNUSED(aclk_hostname);
+    UNUSED(aclk_port);
+    UNUSED(on_connect);
+    UNUSED(on_disconnect);
     return 0;
 }
 
 #else
-/*
- * Just report the library info in the logfile for reference when issues arise
- *
- */
 
 struct mosquitto *mosq = NULL;
 
 // Get a string description of the error
-
 inline const char *_link_strerror(int rc)
 {
     return mosquitto_strerror(rc);
 }
 
-
-void mqtt_message_callback(
-    struct mosquitto *mosq, void *obj, const struct mosquitto_message *msg)
+void mqtt_message_callback(struct mosquitto *mosq, void *obj, const struct mosquitto_message *msg)
 {
-    (void) mosq;
-    (void) obj;
-
-    // TODO: handle commands in a more efficient way, if we have many
-
-    if (strcmp((char *)msg->payload, "pause") == 0) {
-        cmdpause = 1;
-        return;
-    }
-
-    if (strcmp((char *)msg->payload, "resume") == 0) {
-        cmdpause = 0;
-        return;
-    }
-
-    if (strcmp((char *)msg->payload, "reload") == 0) {
-        error_log_limit_unlimited();
-        info("Reloading health configuration");
-        health_reload();
-        error_log_limit_reset();
-        return;
-    }
-
-    if (strcmp((char *)msg->payload, "info") == 0) {
-        aclk_send_metadata();
-        return;
-    }
+    UNUSED(mosq);
+    UNUSED(obj);
 
     aclk_handle_cloud_request(msg->payload);
-
-    //info("Received type=[%s], msg-id=[%s], topic=[%s], url=[%s]",cloud_to_agent.type_id, cloud_to_agent.msg_id, cloud_to_agent.topic, cloud_to_agent.url);
-
 }
-
-int lws_wss_client_initialized = 0;
 
 // This is not define because in future we might want to try plain
 // MQTT as fallback ?
@@ -113,47 +75,53 @@ int lws_wss_client_initialized = 0;
 int mqtt_over_websockets = 1;
 struct aclk_lws_wss_engine_instance *lws_engine_instance = NULL;
 
+void publish_callback(struct mosquitto *mosq, void *obj, int rc)
+{
+    UNUSED(mosq);
+    UNUSED(obj);
+    UNUSED(rc);
+
+    // TODO: link this with a msg_id so it can be traced
+    return;
+}
+
 void connect_callback(struct mosquitto *mosq, void *obj, int rc)
 {
-    (void) obj;
-    (void) rc;
+    UNUSED(obj);
+    UNUSED(rc);
 
     info("Connection to cloud estabilished");
 
-    aclk_connection_initialized = 1;
     aclk_mqtt_connected = 1;
-    _on_connect((void *) mosq);
+    _on_connect((void *)mosq);
 
     return;
 }
-
 
 void disconnect_callback(struct mosquitto *mosq, void *obj, int rc)
 {
-    (void) obj;
-    (void) rc;
+    UNUSED(obj);
+    UNUSED(rc);
 
     info("Connection to cloud failed");
-    // TODO: Keep the connection "alive" for now. The library will reconnect.
 
-    //mqtt_connection_initialized = 0;
     aclk_mqtt_connected = 0;
-    _on_disconnect((void *) mosq);
+    _on_disconnect((void *)mosq);
 
-    if(mqtt_over_websockets && lws_engine_instance)
+    if (mqtt_over_websockets && lws_engine_instance)
         aclk_lws_wss_mqtt_layer_disconect_notif(lws_engine_instance);
 
-    //sleep_usec(USEC_PER_SEC * 5);
     return;
 }
-
 
 void _show_mqtt_info()
 {
     int libmosq_major, libmosq_minor, libmosq_revision, libmosq_version;
-    libmosq_version =  mosquitto_lib_version(&libmosq_major, &libmosq_minor, &libmosq_revision);
+    libmosq_version = mosquitto_lib_version(&libmosq_major, &libmosq_minor, &libmosq_revision);
 
-    info("Detected libmosquitto library version %d, %d.%d.%d",libmosq_version, libmosq_major, libmosq_minor, libmosq_revision);
+    info(
+        "Detected libmosquitto library version %d, %d.%d.%d", libmosq_version, libmosq_major, libmosq_minor,
+        libmosq_revision);
 }
 
 size_t _mqtt_external_write_hook(void *buf, size_t count)
@@ -166,7 +134,7 @@ size_t _mqtt_external_read_hook(void *buf, size_t count)
     return aclk_lws_wss_client_read(lws_engine_instance, buf, count);
 }
 
-int _mqtt_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *), void (*on_disconnect)(void *))
+int _mqtt_lib_init(void (*on_connect)(void *), void (*on_disconnect)(void *))
 {
     int rc;
     int libmosq_major, libmosq_minor, libmosq_revision, libmosq_version;
@@ -174,12 +142,11 @@ int _mqtt_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *
     char *server_crt;
     char *server_key;
 
-    // show library info so can have in in the logfile
+    // show library info so can have it in the logfile
     libmosq_version = mosquitto_lib_version(&libmosq_major, &libmosq_minor, &libmosq_revision);
     ca_crt = config_get(CONFIG_SECTION_ACLK, "agent cloud link cert", "*");
     server_crt = config_get(CONFIG_SECTION_ACLK, "agent cloud link server cert", "*");
     server_key = config_get(CONFIG_SECTION_ACLK, "agent cloud link server key", "*");
-
 
     if (ca_crt[0] == '*') {
         freez(ca_crt);
@@ -196,9 +163,9 @@ int _mqtt_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *
         server_key = NULL;
     }
 
-    info(
-        "Detected libmosquitto library version %d, %d.%d.%d", libmosq_version, libmosq_major, libmosq_minor,
-        libmosq_revision);
+    //    info(
+    //        "Detected libmosquitto library version %d, %d.%d.%d", libmosq_version, libmosq_major, libmosq_minor,
+    //        libmosq_revision);
 
     rc = mosquitto_lib_init();
     if (unlikely(rc != MOSQ_ERR_SUCCESS)) {
@@ -218,6 +185,7 @@ int _mqtt_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *
 
     mosquitto_connect_callback_set(mosq, connect_callback);
     mosquitto_disconnect_callback_set(mosq, disconnect_callback);
+    mosquitto_publish_callback_set(mosq, publish_callback);
 
     mosquitto_username_pw_set(mosq, NULL, NULL);
 
@@ -234,8 +202,8 @@ int _mqtt_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *
     info("MQTT in flight messages set to 1  -- %s", mosquitto_strerror(rc));
 #endif
 
-    if(!mqtt_over_websockets) {
-        rc = mosquitto_reconnect_delay_set(mosq, ACLK_RECONNECT_DELAY, ACLK_MAX_RECONNECT_DELAY, 1);
+    if (!mqtt_over_websockets) {
+        rc = mosquitto_reconnect_delay_set(mosq, ACLK_RECONNECT_DELAY, ACLK_MAX_BACKOFF_DELAY, 1);
 
         if (unlikely(rc != MOSQ_ERR_SUCCESS))
             error("Failed to set the reconnect delay (%d) (%s)", rc, mosquitto_strerror(rc));
@@ -253,9 +221,11 @@ int _link_mqtt_connect(char *aclk_hostname, int aclk_port)
     rc = mosquitto_connect_async(mosq, aclk_hostname, aclk_port, ACLK_PING_INTERVAL);
 
     if (unlikely(rc != MOSQ_ERR_SUCCESS))
-        error("Connect %s MQTT status = %d (%s)", aclk_hostname, rc, mosquitto_strerror(rc));
+        error(
+            "Failed to establish link to [%s:%d] MQTT status = %d (%s)", aclk_hostname, aclk_port, rc,
+            mosquitto_strerror(rc));
     else
-        info("Establishing MQTT link to %s", aclk_hostname);
+        info("Establishing MQTT link to [%s:%d]", aclk_hostname, aclk_port);
 
     return rc;
 }
@@ -264,60 +234,72 @@ static inline void _link_mosquitto_write()
 {
     int rc;
 
-    if(!mqtt_over_websockets)
+    if (!mqtt_over_websockets)
         return;
 
     rc = mosquitto_loop_misc(mosq);
-    if(unlikely( rc != MOSQ_ERR_SUCCESS ))
+    if (unlikely(rc != MOSQ_ERR_SUCCESS))
         debug(D_ACLK, "ACLK: failure during mosquitto_loop_misc %s", mosquitto_strerror(rc));
 
-    if(likely( mosquitto_want_write(mosq) )) {
+    if (likely(mosquitto_want_write(mosq))) {
         rc = mosquitto_loop_write(mosq, 1);
-        if( rc != MOSQ_ERR_SUCCESS )
+        if (rc != MOSQ_ERR_SUCCESS)
             debug(D_ACLK, "ACLK: failure during mosquitto_loop_write %s", mosquitto_strerror(rc));
     }
 }
 
-void aclk_lws_connect_notif_callback(){
+void aclk_lws_connect_notif_callback()
+{
     //the connection is done by LWS so this parameters dont matter
     //ig MQTT over LWS is used
-    _link_mqtt_connect("doesntmatter", 12345);
+    _link_mqtt_connect(aclk_hostname, aclk_port);
     _link_mosquitto_write();
 }
 
-void aclk_lws_data_received_callback(){
+void aclk_lws_data_received_callback()
+{
     int rc = mosquitto_loop_read(mosq, 1);
-    if(rc != MOSQ_ERR_SUCCESS)
-        debug(D_ACLK, "ACLK: failure during mosquitto_loop_read %s", mosquitto_strerror(rc));       
+    if (rc != MOSQ_ERR_SUCCESS)
+        debug(D_ACLK, "ACLK: failure during mosquitto_loop_read %s", mosquitto_strerror(rc));
+}
+
+void aclk_lws_connection_closed()
+{
+    aclk_disconnect(NULL);
 }
 
 static const struct aclk_lws_wss_engine_callbacks aclk_lws_engine_callbacks = {
     .connection_established_callback = aclk_lws_connect_notif_callback,
     .data_rcvd_callback = aclk_lws_data_received_callback,
-    .data_writable_callback = NULL
+    .data_writable_callback = NULL,
+    .connection_closed = aclk_lws_connection_closed
 };
 
 int _link_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *), void (*on_disconnect)(void *))
 {
     int rc;
 
-    if(mqtt_over_websockets) {
+    if (mqtt_over_websockets) {
         // we will connect when WebSocket connection is up
         // based on callback
-        if(!lws_wss_client_initialized) {
+        if (!lws_engine_instance)
             lws_engine_instance = aclk_lws_wss_client_init(&aclk_lws_engine_callbacks, aclk_hostname, aclk_port);
-            aclk_lws_wss_service_loop(lws_engine_instance);
-            lws_wss_client_initialized = 1;
-        }
+        else
+            aclk_lws_wss_connect(lws_engine_instance);
+
+        aclk_lws_wss_service_loop(lws_engine_instance);
     }
 
-    rc = _mqtt_lib_init(aclk_hostname, aclk_port, on_connect, on_disconnect);
-    if(rc != MOSQ_ERR_SUCCESS)
+    rc = _mqtt_lib_init(on_connect, on_disconnect);
+    if (rc != MOSQ_ERR_SUCCESS)
         return rc;
 
-    if(mqtt_over_websockets) {
+    if (mqtt_over_websockets) {
         mosquitto_external_callbacks_set(mosq, _mqtt_external_write_hook, _mqtt_external_read_hook);
-        return MOSQ_ERR_SUCCESS;
+        if (!lws_engine_instance)
+            return 1;
+        else
+            return MOSQ_ERR_SUCCESS;
     } else {
         // if direct mqtt connection is used
         // connect immediatelly
@@ -327,7 +309,11 @@ int _link_lib_init(char *aclk_hostname, int aclk_port, void (*on_connect)(void *
 
 static inline int _link_event_loop_wss()
 {
-    if(lws_engine_instance && lws_engine_instance->websocket_connection_up)
+    if (unlikely(!lws_engine_instance)) {
+        return MOSQ_ERR_SUCCESS;
+    }
+
+    if (lws_engine_instance && lws_engine_instance->websocket_connection_up)
         _link_mosquitto_write();
 
     aclk_lws_wss_service_loop(lws_engine_instance);
@@ -357,9 +343,9 @@ static inline int _link_event_loop_plain_mqtt(int timeout)
 
 int _link_event_loop(int timeout)
 {
-    if(mqtt_over_websockets)
+    if (mqtt_over_websockets)
         return _link_event_loop_wss();
-    
+
     return _link_event_loop_plain_mqtt(timeout);
 }
 
@@ -380,7 +366,7 @@ void _link_shutdown()
     mosquitto_destroy(mosq);
     mosq = NULL;
 
-    if(lws_engine_instance) {
+    if (lws_engine_instance) {
         aclk_lws_wss_client_destroy(lws_engine_instance);
         lws_engine_instance = NULL;
     }
@@ -388,8 +374,7 @@ void _link_shutdown()
     return;
 }
 
-
-int _link_subscribe(char  *topic, int qos)
+int _link_subscribe(char *topic, int qos)
 {
     int rc;
 
@@ -410,13 +395,12 @@ int _link_subscribe(char  *topic, int qos)
     return 0;
 }
 
-
 /*
  * Send a message to the cloud to specific topic
  *
  */
 
-int _link_send_message(char *topic, char *message)
+int _link_send_message(char *topic, char *message, int *mid)
 {
     int rc;
 
@@ -427,15 +411,11 @@ int _link_send_message(char *topic, char *message)
 
     int msg_len = strlen(message);
 
-    // TODO: handle encoding validation -- the message should be UFT8 encoded by the sender
-    //rc = mosquitto_validate_utf8(message, msg_len);
-    //if (unlikely(rc != MOSQ_ERR_SUCCESS))
-    //    return rc;
-
-    rc = mosquitto_publish(mosq, NULL, topic, msg_len, message, ACLK_QOS, 0);
+    rc = mosquitto_publish(mosq, mid, topic, msg_len, message, ACLK_QOS, 0);
 
     // TODO: Add better handling -- error will flood the logfile here
     if (unlikely(rc != MOSQ_ERR_SUCCESS)) {
+        errno = 0;
         error("MQTT message failed : %s", mosquitto_strerror(rc));
     }